IMPORTANTE: Ejemplo de MATDIALOG

+++ src/app/heroes/pages/agregar/agregar.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs';
+import { ConfirmarComponent } from '../../components/confirmar/confirmar.component';
 import { Heroe, Publisher } from '../../interfaces/heroes.interface';
 import { HeroesService } from '../../services/heroes.service';
 
@@ -41,7 +43,8 @@ export class AgregarComponent implements OnInit {
   constructor(private heroesService: HeroesService,
               private activatedRoute: ActivatedRoute,
               private router: Router,
-              private snackBar: MatSnackBar) { }
+              private snackBar: MatSnackBar,
+              public dialog: MatDialog) { }
 
   ngOnInit(): void {
 
@@ -69,10 +72,20 @@ export class AgregarComponent implements OnInit {
   }
 
   deleteHeroe(): void {
-    this.heroesService.deleteHeroe(this.heroe.id!).subscribe(() => {
-      this.showSnackBar('Héroe eliminado correctamente');
-      this.router.navigate(['/heroes']);
+    const dialog = this.dialog.open(ConfirmarComponent, {
+      width: '300px',
+      data: {...this.heroe} // Para no pasarlo por referencia
     });
+
+    dialog.afterClosed().subscribe(resp => {
+      if (resp) {
+        this.heroesService.deleteHeroe(this.heroe.id!).subscribe(() => {
+          this.showSnackBar('Héroe eliminado correctamente');
+          this.router.navigate(['/heroes']);
+        });
+      } 
+    });
+
   }
 
   showSnackBar(message: string): void {
